Change: The data type of ECosHelperFileInfoType enum is changed from uint32 to uint8 due to blueprint restrictions

+++ CosHelper/Source/CosHelper/Public/CosHelper.h
@@ -6,7 +6,7 @@
 #include "Interfaces/IHttpRequest.h"
 #include "CosHelper.generated.h"
 
-enum class ECosHelperFileInfoType : uint32;
+enum class ECosHelperFileInfoType : uint8;
 
 struct FCosHelperInitializeInfo;
